Implement Move System with Special Moves for All Pokémon
- Added Move struct in Move.hpp to store move name and power.
- Integrated move system with a vector of moves in the Pokémon class.
- Implemented selectAndUseMove(), printAvailableMoves(), selectMove(), and useMove() functions for move selection and execution during battles.
- Refactored Pokémon classes to use the new move system, reducing repetitive battle code.
- Added unique special moves for each Pokémon with distinct effects:
  - Caterpie: "STICKY WEB" reduces opponent's attack power.
  - Charmander: "BLAZING CHARGE" causes recoil damage.
  - Pidgey: "GUST" can end the battle by blowing away the opponent.
  - Pikachu: "THUNDER BOLT" has a chance to miss.
  - Squirtle: "RAPID SPIN" hits multiple times.
  - Zubat: "LEECH LIFE" restores health based on damage dealt.

+++ Pokemon/src/Battle/WildEncounterManager.cpp
@@ -8,7 +8,7 @@ using namespace N_Pokemon;
 using namespace N_Battle;
 
 WildEncounterManager::WildEncounterManager() {
-	srand(time(0));
+	srand(static_cast<unsigned int>(time (0)));
 }
 
 Pokemon* WildEncounterManager::getRandomPokemonFromGrass(const Grass* grass)
